==> src/1.packages_load.R <==
library(vroom)
library(dplyr)
library(tidyr)
library(tibble)
library(stats)
library(ggplot2)
#BiocManager::install("maSigPro")
library(maSigPro)
#BiocManager::install("Trendy")
library(Trendy)
#library(edgeR)
library(reshape2)
#install.packages("doRNG")
library(doRNG)
#install.packages("doParallel")
library(doParallel)
#source("/Users/nirwantandukar/Documents/GitHub/dynGENIE3/dynGENIE3_R_C_wrapper/dynGENIE3.R")
#library(impute)
#library(preprocessCore)
#library(vars)
#BiocManager::install("bnlearn")
library(bnlearn)
library(igraph)
#library(RCy3)
library(devtools)
#install_github("jpvert/tigress")
#library(tigress)
#BiocManager::install("minet")
library(minet)
#library(Rgraphviz)
#install.packages("coda")
library(coda)
#install.packages("rjags")
# devtools::install_url("http://sourceforge.net/projects/mcmc-jags/files/rjags/3/rjags_3-2.tar.gz",
#                       args="--configure-args='--with-jags-include=/Users/casallas/homebrew/opt/jags/include/JAGS        
#                                               --with-jags-lib=/Users/casallas/homebrew/opt/jags/lib'
#                             "
# )
# Install source code for jags:
# https://sourceforge.net/projects/mcmc-jags/
library(rjags)
#install.packages("abn")
#library(abn)
#install.packages("NMF")
library(NMF)
library(data.table)
#install.packages("dbnR")
library(dbnR)
#install.packages("timeSeq")
library(timeSeq)
